requirement.txt
streamlit
pandas
plotly
google-generativeai
requests
numpy
scikit-learn
pathway
matplotlib
